use util methods
