+++ scripts/tasks_retrival/task_retrieval.py
@@ -172,3 +172,20 @@ def get_id_to_symbol_df(list_of_gene_metadata):
     gene_metadata_df = gene_metadata_df.drop_duplicates(subset="query")
     gene_metadata_df.index = gene_metadata_df["query"]
     return gene_metadata_df
+
+
+def report_numerical_task_task(
+    outcomes: pd.Series, main_task_directory: str, task_name: str
+):
+    """
+    prints a short task report.
+
+    Args:
+    ----
+        df (pd.Series): _description_
+        main_task_directory (str): the main folder for the task
+        task_name (str): The task name
+
+    """
+    print(f"Task saved at {main_task_directory}  under {task_name} /\n")
+    print(f"n = {len(outcomes)} mean {outcomes.mean():.2f} sd {outcomes.std():.2f}")
